TARS/pyproject.toml
[project]
name = "TARS"
version = "0.1.0"
description = "TARS AI Assistant"
requires-python = ">=3.13"
dependencies = [
    "google-auth",
    "google-auth-oauthlib",
    "google-api-python-client",
    "langchain",
    "openai",
    "python-dotenv",
    "slack-bolt",
    "pyowm",
    "yfinance",
    "youtube-search",
    "pydantic-settings",
    "fastapi",
    "streamlit",
    "langchain-anthropic",
    "langchain-openai",
    "langchain-google-genai",
    "langchain-community",
    "langgraph",
    "uvicorn",
    "pinecone-client",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "black",
    "isort",
    "mypy",
    "flake8",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend" 